+++ sandbox/include/sandbox/sandbox.h
@@ -7,4 +7,4 @@ application_t *application_create();
 void application_destroy(application_t *application);
 
 b8 on_render(void);
-b8 on_update(u64 timestep);
+b8 on_update(f64 timestep);
